improve separation of concerns
