+++ src/main/java/com/github/cameltooling/lsp/internal/instancemodel/ReferenceUtils.java
@@ -12,20 +12,33 @@ package com.github.cameltooling.lsp.internal.instancemodel;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Set;
 
 /**
  * @author lheinema
  *
  */
-public class ModelUtils {
+public class ReferenceUtils {
 
 	private static final List<String> POSSIBLE_REFERENCES = Arrays.asList("direct","direct-vm", "seda", "vm");
 
-	private ModelUtils() {
+	private ReferenceUtils() {
 		// util class
 	}
 	
 	public static boolean isReferenceComponentKind(CamelUriElementInstance camelURIInstanceToSearchReference) {
 		return POSSIBLE_REFERENCES.contains(camelURIInstanceToSearchReference.getComponentName());
+	}
+	
+	/**
+	 * @param camelDirectURIInstance
+	 * @return the first path parameter which is the reference key for elements listed in com.github.cameltooling.lsp.internal.instancemodel.ReferenceUtils.POSSIBLE_REFERENCES
+	 */
+	public static String getReferenceKey(CamelURIInstance camelDirectURIInstance) {
+		Set<PathParamURIInstance> pathParams = camelDirectURIInstance.getComponentAndPathUriElementInstance().getPathParams();
+		if (!pathParams.isEmpty()) {
+			return pathParams.iterator().next().getValue();
+		}
+		return null;
 	}
 }
